requirements.txt
phidata
groq 
Ollama
google.genai
psaw
